Refactor the disassembler out of polkatool and into a dedicated crate to make it re-usable
