+++ tools/polkatool/src/main.rs
@@ -4,19 +4,9 @@
 
 use clap::Parser;
 use polkavm_common::program::{Opcode, ProgramBlob};
+use polkavm_disassembler::DisassemblyFormat;
 use std::collections::HashMap;
-use std::{
-    io::Write,
-    path::{Path, PathBuf},
-};
-
-#[derive(Copy, Clone, Debug, clap::ValueEnum)]
-enum DisassemblyFormat {
-    Guest,
-    GuestAndNative,
-    Native,
-    DiffFriendly,
-}
+use std::path::{Path, PathBuf};
 
 #[derive(Parser, Debug)]
 #[clap(version)]
@@ -175,81 +165,10 @@ fn main_stats(inputs: Vec<PathBuf>) -> Result<(), String> {
 fn main_disassemble(input: PathBuf, format: DisassemblyFormat, display_gas: bool, output: Option<PathBuf>) -> Result<(), String> {
     let blob = load_blob(&input)?;
 
-    let native = if matches!(format, DisassemblyFormat::Native | DisassemblyFormat::GuestAndNative) {
-        if !cfg!(target_arch = "x86_64") {
-            bail!("the selected disassembly format is not supported on this architecture");
-        }
-
-        let mut config = match polkavm::Config::from_env() {
-            Ok(config) => config,
-            Err(error) => bail!("failed to fetch VM configuration from the environment: {error}"),
-        };
-
-        config.set_worker_count(0);
-
-        let engine = match polkavm::Engine::new(&config) {
-            Ok(engine) => engine,
-            Err(error) => bail!("failed to create VM engine: {error}"),
-        };
-
-        let module = match polkavm::Module::from_blob(&engine, &Default::default(), &blob) {
-            Ok(module) => module,
-            Err(error) => bail!("failed to compile {input:?}: {error}"),
-        };
-
-        let machine_code = match module.machine_code() {
-            Some(code) => code.into_owned(),
-            None => bail!("currently selected VM backend doesn't provide raw machine code"),
-        };
-
-        let instruction_map = match module.code_offset_to_native_code_offset() {
-            Some(map) => map.to_vec(),
-            None => bail!("currently selected VM backend doesn't provide a machine code map"),
-        };
-
-        let machine_code_origin = module.machine_code_origin().unwrap_or(0);
-        Some((machine_code_origin, machine_code, instruction_map))
-    } else {
-        None
-    };
-
-    let mut gas_cost_map = HashMap::new();
+    let mut disassembler = polkavm_disassembler::Disassembler::new(&blob, format).map_err(|error| error.to_string())?;
     if display_gas {
-        let mut config = match polkavm::Config::from_env() {
-            Ok(config) => config,
-            Err(error) => bail!("failed to fetch VM configuration from the environment: {error}"),
-        };
-
-        config.set_worker_count(0);
-        config.set_backend(Some(polkavm::BackendKind::Interpreter));
-
-        let engine = match polkavm::Engine::new(&config) {
-            Ok(engine) => engine,
-            Err(error) => bail!("failed to create VM engine: {error}"),
-        };
-
-        let mut config = polkavm::ModuleConfig::default();
-        config.set_gas_metering(Some(polkavm::GasMeteringKind::Sync));
-
-        let module = match polkavm::Module::from_blob(&engine, &config, &blob) {
-            Ok(module) => module,
-            Err(error) => bail!("failed to compile {input:?}: {error}"),
-        };
-
-        let mut in_new_block = true;
-        for instruction in blob.instructions() {
-            if in_new_block {
-                in_new_block = false;
-                if let Some(cost) = module.gas_cost_for_code_offset(instruction.offset) {
-                    gas_cost_map.insert(instruction.offset, cost);
-                }
-            }
-
-            if instruction.starts_new_basic_block() {
-                in_new_block = true;
-            }
-        }
-    };
+        disassembler.display_gas().map_err(|error| error.to_string())?;
+    }
 
     match output {
         Some(output) => {
@@ -260,286 +179,12 @@ fn main_disassemble(input: PathBuf, format: DisassemblyFormat, display_gas: bool
                 }
             };
 
-            disassemble_into(format, &blob, native, gas_cost_map, std::io::BufWriter::new(fp))
+            disassembler.disassemble_into(std::io::BufWriter::new(fp))
         }
         None => {
             let stdout = std::io::stdout();
-            disassemble_into(format, &blob, native, gas_cost_map, std::io::BufWriter::new(stdout))
+            disassembler.disassemble_into(std::io::BufWriter::new(stdout))
         }
     }
-}
-
-#[derive(Default)]
-struct AssemblyFormatter {
-    buffer: String,
-}
-
-impl AssemblyFormatter {
-    fn emit(
-        &mut self,
-        indent: bool,
-        code_origin: u64,
-        mut code: &[u8],
-        mut position: usize,
-        writer: &mut impl Write,
-    ) -> Result<(), std::io::Error> {
-        use iced_x86::Formatter;
-
-        let mut formatter = iced_x86::NasmFormatter::new();
-        formatter.options_mut().set_space_after_operand_separator(true);
-        formatter.options_mut().set_hex_prefix("0x");
-        formatter.options_mut().set_hex_suffix("");
-        formatter.options_mut().set_uppercase_hex(false);
-        formatter.options_mut().set_small_hex_numbers_in_decimal(false);
-        formatter.options_mut().set_show_useless_prefixes(true);
-        formatter.options_mut().set_branch_leading_zeros(false);
-        formatter.options_mut().set_rip_relative_addresses(true);
-
-        loop {
-            let mut decoder = iced_x86::Decoder::with_ip(64, code, code_origin, iced_x86::DecoderOptions::NONE);
-            if !decoder.can_decode() {
-                break;
-            }
-            let mut instruction = iced_x86::Instruction::default();
-            decoder.decode_out(&mut instruction);
-
-            if indent {
-                write!(writer, "                                       ")?;
-            }
-            write!(writer, "{:8x}: ", position as u64 + code_origin)?;
-
-            let start_index = (instruction.ip() - code_origin) as usize;
-            let instr_bytes = &code[start_index..start_index + instruction.len()];
-            let mut count = 0;
-            for b in instr_bytes.iter() {
-                write!(writer, "{:02x} ", b)?;
-                count += 3;
-            }
-            while count < 34 {
-                write!(writer, " ")?;
-                count += 1;
-            }
-
-            self.buffer.clear();
-            formatter.format(&instruction, &mut self.buffer);
-            write!(writer, "{}", self.buffer)?;
-            writeln!(writer)?;
-
-            code = &code[instruction.len()..];
-            position += instruction.len();
-        }
-
-        Ok(())
-    }
-}
-
-#[allow(clippy::type_complexity)]
-fn disassemble_into(
-    format: DisassemblyFormat,
-    blob: &polkavm_linker::ProgramBlob,
-    native: Option<(u64, Vec<u8>, Vec<(u32, u32)>)>,
-    gas_cost_map: HashMap<u32, i64>,
-    mut writer: impl Write,
-) -> Result<(), String> {
-    let mut instructions = Vec::new();
-    for instruction in blob.instructions() {
-        instructions.push((instruction.offset, instruction.kind));
-    }
-
-    let mut exports_for_code_offset = HashMap::new();
-    for (nth_export, export) in blob.exports().enumerate() {
-        exports_for_code_offset
-            .entry(export.target_code_offset())
-            .or_insert_with(Vec::new)
-            .push((nth_export, export));
-    }
-
-    let mut jump_table_map = HashMap::new();
-    let mut jump_table = Vec::new();
-    for target_code_offset in blob.jump_table() {
-        let jump_table_index = jump_table.len() + 1;
-        jump_table.push(target_code_offset);
-        assert!(jump_table_map.insert(target_code_offset, jump_table_index).is_none());
-    }
-
-    let format_jump_target = |target_offset: u32, basic_block_counter: u32| {
-        use core::fmt::Write;
-
-        let mut buf = String::new();
-        if !matches!(format, DisassemblyFormat::DiffFriendly) {
-            write!(&mut buf, "@{basic_block_counter}").unwrap()
-        } else {
-            buf.push_str("@_:");
-        }
-
-        if let Some(jump_table_index) = jump_table_map.get(&target_offset) {
-            if !matches!(format, DisassemblyFormat::DiffFriendly) {
-                write!(&mut buf, " [@dyn {jump_table_index}]").unwrap()
-            } else {
-                buf.push_str(" [_]");
-            }
-        }
-
-        if let Some(exports) = exports_for_code_offset.get(&target_offset) {
-            for (nth_export, export) in exports {
-                write!(&mut buf, " [export #{}: {}]", nth_export, export.symbol()).unwrap()
-            }
-        }
-
-        if let Some(gas_cost) = gas_cost_map.get(&target_offset) {
-            write!(&mut buf, " (gas: {})", gas_cost).unwrap();
-        }
-
-        buf
-    };
-
-    let mut fmt = AssemblyFormatter::default();
-    let mut last_line_program_entry = None;
-    let mut last_full_name = String::new();
-    let mut basic_block_counter = 0;
-    let mut pending_label = true;
-    for (nth_instruction, (offset, instruction)) in instructions.iter().copied().enumerate() {
-        let instruction_s = if let polkavm_common::program::Instruction::ecalli(nth_import) = instruction {
-            if let Some(import) = blob.imports().get(nth_import) {
-                format!("{instruction} // {}", import)
-            } else {
-                format!("{instruction} // INVALID")
-            }
-        } else {
-            instruction.to_string()
-        };
-
-        let line_program = match blob.get_debug_line_program_at(nth_instruction as u32) {
-            Ok(line_program) => line_program,
-            Err(error) => {
-                bail!("failed to parse line program: {error}");
-            }
-        };
-
-        if let Some(mut line_program) = line_program {
-            if last_line_program_entry != Some(line_program.entry_index()) {
-                if nth_instruction != 0 {
-                    if let Err(error) = writeln!(&mut writer) {
-                        bail!("failed to write to output: {error}");
-                    }
-                }
-
-                last_line_program_entry = Some(line_program.entry_index());
-                loop {
-                    let region = match line_program.run() {
-                        Ok(Some(region)) => region,
-                        Ok(None) => break,
-                        Err(error) => {
-                            bail!("failed to parse line program: {error}");
-                        }
-                    };
-
-                    if region.instruction_range().contains(&(nth_instruction as u32)) {
-                        let frame = region.frames().next().unwrap();
-                        let full_name = match frame.full_name() {
-                            Ok(full_name) => full_name,
-                            Err(error) => {
-                                bail!("failed to parse line program: {error}");
-                            }
-                        }
-                        .to_string();
-
-                        if last_full_name != full_name {
-                            if let Err(error) = writeln!(&mut writer, "<{}>:", full_name) {
-                                bail!("failed to write to output: {error}");
-                            }
-
-                            last_full_name = full_name;
-                        }
-
-                        break;
-                    }
-                }
-            }
-        } else {
-            if !last_full_name.is_empty() {
-                if let Err(error) = writeln!(&mut writer) {
-                    bail!("failed to write to output: {error}");
-                }
-            }
-
-            last_line_program_entry = None;
-            last_full_name.clear();
-        }
-
-        if pending_label {
-            pending_label = false;
-            let result = if !matches!(format, DisassemblyFormat::DiffFriendly) {
-                writeln!(&mut writer, "      : {}", format_jump_target(offset, basic_block_counter))
-            } else {
-                writeln!(&mut writer, "    {}", format_jump_target(offset, basic_block_counter))
-            };
-
-            if let Err(error) = result {
-                bail!("failed to write to output: {error}");
-            }
-        }
-
-        if matches!(format, DisassemblyFormat::DiffFriendly) {
-            let mut string = instruction_s;
-            if let polkavm_common::program::Instruction::load_imm(dst, _) = instruction {
-                string = format!("{} = _", dst);
-            }
-
-            if let Some(index) = string.find('@') {
-                let length = string[index + 1..]
-                    .chars()
-                    .take_while(|character| character.is_ascii_digit() || matches!(character, 'a' | 'b' | 'c' | 'd' | 'e' | 'f'))
-                    .count();
-                string.replace_range(index + 1..index + 1 + length, "_");
-            }
-
-            if let Some(index_1) = string.find("[0x") {
-                let index_2 = string[index_1..].find(']').unwrap() + index_1;
-                string.replace_range(index_1..=index_2, "[_]");
-            }
-
-            if let Err(error) = writeln!(&mut writer, "    {}", string) {
-                bail!("failed to write to output: {error}");
-            }
-        } else if matches!(format, DisassemblyFormat::Guest | DisassemblyFormat::GuestAndNative) {
-            if let Err(error) = writeln!(&mut writer, "{offset:6}: {instruction_s}") {
-                bail!("failed to write to output: {error}");
-            }
-        }
-
-        if matches!(format, DisassemblyFormat::Native | DisassemblyFormat::GuestAndNative) {
-            let (machine_code_origin, machine_code, map) = native.as_ref().unwrap();
-            assert_eq!(offset, map[nth_instruction].0);
-
-            let machine_code_position = map[nth_instruction].1 as usize;
-            let machine_next_code_position = map[nth_instruction + 1].1 as usize;
-            let length = machine_next_code_position - machine_code_position;
-            if length != 0 {
-                let machine_code_chunk = &machine_code[machine_code_position..machine_next_code_position];
-                if let Err(error) = fmt.emit(
-                    matches!(format, DisassemblyFormat::GuestAndNative),
-                    *machine_code_origin,
-                    machine_code_chunk,
-                    machine_code_position,
-                    &mut writer,
-                ) {
-                    bail!("failed to write to output: {error}");
-                }
-            }
-        }
-
-        if instruction.opcode().starts_new_basic_block() {
-            if nth_instruction + 1 != instructions.len() {
-                pending_label = true;
-            }
-            basic_block_counter += 1;
-        }
-    }
-
-    if let Err(error) = writer.flush() {
-        bail!("failed to write to output: {error}");
-    }
-
-    Ok(())
+    .map_err(|error| error.to_string())
 }
